undid stick keys

+++ src/ie/dit/main/BasicEnemyTrail.java
@@ -17,7 +17,7 @@ public class BasicEnemyTrail extends GameObject {
 	private int height;
 	
 	
-	public BasicEnemyTrail(int x, int y, ID id,Color color ,int width ,int height,float life,Handler handler) {
+	public BasicEnemyTrail(float x, float y, ID id,Color color ,int width ,int height,float life,Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		this.color = color;
@@ -49,7 +49,7 @@ public class BasicEnemyTrail extends GameObject {
 		Graphics2D g2d = (Graphics2D) g;
 		g2d.setComposite(makeTransparent(alpha));
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x,(int) y, width, height);
 		
 		g2d.setComposite(makeTransparent(1)); //fixes bug which was making random things transparent
 	}
