+++ src/ie/dit/main/GameObject.java
@@ -5,11 +5,12 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velX,velY;
+	protected float velX,velY;
+
 	
-	public GameObject(int x, int y ,ID id)
+	public GameObject(float x, float y ,ID id)
 	{
 		this.x = x;
 		this.y = y;
@@ -34,11 +35,11 @@ public abstract class GameObject {
 	}
 	
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
@@ -63,18 +64,14 @@ public abstract class GameObject {
 		this.velY = velY;
 	}
 	
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
 
-	public int getVelY() {
+	public float getVelY() {
 		return velY;
 	}
 
-	
-	
-	
-	
 	
 }
